requirements.txt
langgraph
crewai
openai
langchain
langchain-community
langchain-openai
tavily-python
pydantic
litellm
python-dotenv